+++ client/my-sites/checkout/composite-checkout/wpcom/components/wp-checkout.js
@@ -6,7 +6,6 @@ import { useTranslate } from 'i18n-calypso';
 import styled from '@emotion/styled';
 import {
 	Checkout,
-	CheckoutStepBody,
 	CheckoutSteps,
 	CheckoutStep,
 	getDefaultPaymentMethodStep,
@@ -25,7 +24,7 @@ import {
 import { areDomainsInLineItems, isLineItemADomain } from '../hooks/has-domains';
 import useCouponFieldState from '../hooks/use-coupon-field-state';
 import WPCheckoutOrderReview from './wp-checkout-order-review';
-import WPCheckoutOrderSummary, { WPCheckoutOrderSummaryTitle } from './wp-checkout-order-summary';
+import WPCheckoutOrderSummary from './wp-checkout-order-summary';
 import WPContactForm from './wp-contact-form';
 import { isCompleteAndValid } from '../types';
 import { WPOrderReviewTotal, WPOrderReviewSection, LineItemUI } from './wp-order-review-line-items';
@@ -129,94 +128,89 @@ export default function WPCheckout( {
 	};
 
 	return (
-		<Checkout>
-			<CheckoutStepBody
-				activeStepContent={ null }
-				completeStepContent={ <WPCheckoutOrderSummary siteUrl={ siteUrl } /> }
-				titleContent={ <WPCheckoutOrderSummaryTitle /> }
-				errorMessage={ translate( 'There was an error with the summary step.' ) }
-				isStepActive={ false }
-				isStepComplete={ true }
-				stepNumber={ 1 }
-				totalSteps={ 1 }
-				stepId={ 'order-summary' }
-			/>
-			<CheckoutSteps>
-				<CheckoutStep
-					stepId="review-order-step"
-					isCompleteCallback={ () => true }
-					activeStepContent={
-						<WPCheckoutOrderReview
-							removeItem={ removeItem }
-							couponStatus={ couponStatus }
-							couponFieldStateProps={ couponFieldStateProps }
-							removeCoupon={ removeCouponAndResetActiveStep }
-							onChangePlanLength={ changePlanLength }
-							variantRequestStatus={ variantRequestStatus }
-							variantSelectOverride={ variantSelectOverride }
-							getItemVariants={ getItemVariants }
-						/>
-					}
-					titleContent={ <OrderReviewTitle /> }
-					completeStepContent={ <InactiveOrderReview /> }
-					editButtonText={ translate( 'Edit' ) }
-					editButtonAriaLabel={ translate( 'Edit the payment method' ) }
-					nextStepButtonText={ translate( 'Continue' ) }
-					nextStepButtonAriaLabel={ translate( 'Continue with the selected payment method' ) }
-					validatingButtonText={ translate( 'Please wait…' ) }
-					validatingButtonAriaLabel={ translate( 'Please wait…' ) }
-				/>
-				{ shouldShowContactStep && (
+		<>
+			<WPCheckoutOrderSummary />
+			<Checkout>
+				<CheckoutSteps>
 					<CheckoutStep
-						stepId={ 'contact-form' }
-						isCompleteCallback={ () => {
-							setShouldShowContactDetailsValidationErrors( true );
-							return contactValidationCallback();
-						} }
+						stepId="review-order-step"
+						isCompleteCallback={ () => true }
 						activeStepContent={
-							<WPContactForm
+							<WPCheckoutOrderReview
+								removeItem={ removeItem }
+								couponStatus={ couponStatus }
+								couponFieldStateProps={ couponFieldStateProps }
+								removeCoupon={ removeCouponAndResetActiveStep }
+								onChangePlanLength={ changePlanLength }
+								variantRequestStatus={ variantRequestStatus }
+								variantSelectOverride={ variantSelectOverride }
+								getItemVariants={ getItemVariants }
 								siteUrl={ siteUrl }
-								isComplete={ false }
-								isActive={ true }
-								CountrySelectMenu={ CountrySelectMenu }
-								countriesList={ countriesList }
-								StateSelect={ StateSelect }
-								renderDomainContactFields={ renderDomainContactFields }
-								shouldShowContactDetailsValidationErrors={
-									shouldShowContactDetailsValidationErrors
-								}
 							/>
 						}
-						completeStepContent={ <WPContactForm summary isComplete={ true } isActive={ false } /> }
-						titleContent={ <ContactFormTitle /> }
+						titleContent={ <OrderReviewTitle /> }
+						completeStepContent={ <InactiveOrderReview /> }
 						editButtonText={ translate( 'Edit' ) }
-						editButtonAriaLabel={ translate( 'Edit the contact details' ) }
+						editButtonAriaLabel={ translate( 'Edit the payment method' ) }
 						nextStepButtonText={ translate( 'Continue' ) }
-						nextStepButtonAriaLabel={ translate( 'Continue with the entered contact details' ) }
+						nextStepButtonAriaLabel={ translate( 'Continue with the selected payment method' ) }
 						validatingButtonText={ translate( 'Please wait…' ) }
 						validatingButtonAriaLabel={ translate( 'Please wait…' ) }
 					/>
-				) }
-				<CheckoutStep
-					stepId="payment-method-step"
-					activeStepContent={
-						<PaymentMethodStep
-							CheckoutTerms={ CheckoutTerms }
-							responseCart={ responseCart }
-							subtotal={ subtotal }
+					{ shouldShowContactStep && (
+						<CheckoutStep
+							stepId={ 'contact-form' }
+							isCompleteCallback={ () => {
+								setShouldShowContactDetailsValidationErrors( true );
+								return contactValidationCallback();
+							} }
+							activeStepContent={
+								<WPContactForm
+									siteUrl={ siteUrl }
+									isComplete={ false }
+									isActive={ true }
+									CountrySelectMenu={ CountrySelectMenu }
+									countriesList={ countriesList }
+									StateSelect={ StateSelect }
+									renderDomainContactFields={ renderDomainContactFields }
+									shouldShowContactDetailsValidationErrors={
+										shouldShowContactDetailsValidationErrors
+									}
+								/>
+							}
+							completeStepContent={
+								<WPContactForm summary isComplete={ true } isActive={ false } />
+							}
+							titleContent={ <ContactFormTitle /> }
+							editButtonText={ translate( 'Edit' ) }
+							editButtonAriaLabel={ translate( 'Edit the contact details' ) }
+							nextStepButtonText={ translate( 'Continue' ) }
+							nextStepButtonAriaLabel={ translate( 'Continue with the entered contact details' ) }
+							validatingButtonText={ translate( 'Please wait…' ) }
+							validatingButtonAriaLabel={ translate( 'Please wait…' ) }
 						/>
-					}
-					completeStepContent={ paymentMethodStep.completeStepContent }
-					titleContent={ paymentMethodStep.titleContent }
-					editButtonText={ translate( 'Edit' ) }
-					editButtonAriaLabel={ translate( 'Edit the payment method' ) }
-					nextStepButtonText={ translate( 'Continue' ) }
-					nextStepButtonAriaLabel={ translate( 'Continue with the selected payment method' ) }
-					validatingButtonText={ translate( 'Please wait…' ) }
-					validatingButtonAriaLabel={ translate( 'Please wait…' ) }
-				/>
-			</CheckoutSteps>
-		</Checkout>
+					) }
+					<CheckoutStep
+						stepId="payment-method-step"
+						activeStepContent={
+							<PaymentMethodStep
+								CheckoutTerms={ CheckoutTerms }
+								responseCart={ responseCart }
+								subtotal={ subtotal }
+							/>
+						}
+						completeStepContent={ paymentMethodStep.completeStepContent }
+						titleContent={ paymentMethodStep.titleContent }
+						editButtonText={ translate( 'Edit' ) }
+						editButtonAriaLabel={ translate( 'Edit the payment method' ) }
+						nextStepButtonText={ translate( 'Continue' ) }
+						nextStepButtonAriaLabel={ translate( 'Continue with the selected payment method' ) }
+						validatingButtonText={ translate( 'Please wait…' ) }
+						validatingButtonAriaLabel={ translate( 'Please wait…' ) }
+					/>
+				</CheckoutSteps>
+			</Checkout>
+		</>
 	);
 }
 
@@ -228,7 +222,7 @@ function PaymentMethodStep( { CheckoutTerms, responseCart, subtotal } ) {
 	const [ items, total ] = useLineItems();
 	const taxes = items.filter( ( item ) => item.type === 'tax' );
 	return (
-		<React.Fragment>
+		<>
 			{ paymentMethodStep.activeStepContent }
 
 			<CheckoutTermsUI>
@@ -242,7 +236,7 @@ function PaymentMethodStep( { CheckoutTerms, responseCart, subtotal } ) {
 				) ) }
 				<WPOrderReviewTotal total={ total } />
 			</WPOrderReviewSection>
-		</React.Fragment>
+		</>
 	);
 }
